#9 | Moved some code to the apistore
